requirements.txt
python-telegram-bot==8.0.0